Changes for version 1.2.4
Changes for report.php -- teachers only see groups they are allowed to see and the groups are in a dropdown menu. The other setting options are in toggle buttons. Setting preferences are in hidden values in the button forms. All help or information strings are in tooltips. The buttons are generated through a function, liveview_button. The compact view is supported and truncated question texts are shown in tooltips. the buttons are displayed in a table.