Singleq and other changes
Version 1.2.5 -- Buttons and tooltips for all links/options. Singleqid is set >0 if only one question is desired. this shows the report page with only one question. If the question is multichoice or similar, a histogram is displayed. The singleqid view has all the buttons of the regular page (except for compact the table) and has a button so that the teacher can select another question. The groups are listed alphabetically. Most of the functions have been moved to a new file, the locallib.php file. some new language strings have been added.