Added feature to show students who have not answered
For the single question view, if names are hidden, there is now an option to click and find the names of the students who have not yet answered that question. Also, the single question view goes almost full screen. I also updated a few of the language strings and changed the version number.