Refresh button, session, and immediatefeedback
I removed debug statements that were in the prior version.
I changed the refresh text to a button.
I added in session values so that the options selected would follow the teacher from quiz to quiz.
I changed the code in locallib to handle quizzes set to immediatefeedback.
I changed the version numer and added changes to README.txt