matrix, attachments, and active
The code has been changed so that
A. The program handles matrix questions correctly, listing their choices as the answer.
B. The program handles correctly attachments that students submit as part of an answer to a question, providing a clickable icon for the attachment as part of their answer.
C. The students who have been active in Moodle during the past 5 minutes have a green background for the cell where their name is.