Revise refresh, iframe, pics, text formatting
A major revision (moved to version 3.0.1).
For all questions, the table is now in an iframe. With new responses, the iframe refreshes but not the parent page. For text responses, in single question view, the text response is formatted better. Pictures in the text of the question are displayed. Links are displayed for attachments.