Caught count and question text errors
Under some conditions, the value of $sofar returned from liveview_who_sofar_gridview is not countable. I fixed that.
I changed the variable names in liveviewgrid_display_question so that the displayed question is different from the question object and so that, if the question can't be found, the text of the question is displayed in the old way.